Pass object_name into region_select as string.
This change was done due to object_name being passed in as a symbol if a
symbol was used when using FormHelper. This caused an error in the
set_html_options where the symbol could not be duplicated. This is fixed
by converting the symbol into a string so that duplication can occur in
set_html_options.