Allow prioritization of options and small cleanup
Prioritization of options can now be done by passing in a hash of arrays with
keys corresponding to either :city, :province, or :district. The array
for that key will then determine the order in the choice array so that
items choices appear in order as they were specified. Options that are
outside of the priority options are then included after the prioritized
choices in the default order.