Refactor to regex validation
For some reason and I have no time to fix this or maybe it's a default behaviour
but the form helper get's called twice. Once on render and another on submit or
another loop.

I will take a look at this another time but for now this works and fixes the issue.